pyproject.toml
[tool.poetry]
name = "fds-py"
version = "0.1.0"
description = ""
authors = ["Saikat Karmakar"]
license = "LGPL-3.0, GPL-3.0"
readme = "README.md"
packages = [{ include = "fds", from = "src" }]

[tool.poetry.dependencies]
python = "^3.9"          # ">=3.9,<3.12"
eth-ape = "^0.6.22"
ape-alchemy = "*"
ape-solidity = "*"
ape-foundry = "*"
ape-etherscan = "*"
pyscaffold = "^4.5"
cryptography = "^41.0.4"
# eth-tester = "^0.9.1b1"
coverage-enable-subprocess = "^1.0"
pytest-cov = "^4.1.0"
black = "^23.9.1"
mypy = "^1.5.1"
types-pyyaml = "^6.0.12.12"
types-setuptools = "^68.2.0.0"
types-sqlalchemy = ">=1.4.51"
flake8 = ">=6.1.0,<7"
flake8-breakpoint = ">=1.1.0,<2"
flake8-print = ">=4.0.1,<5"
isort = ">=5.10.1,<6"
mdformat = ">=0.7.17"
mdformat-gfm = ">=0.3.5"
mdformat-frontmatter = ">=0.4.1"
mdformat-pyproject = ">=0.0.1"
pyproject-flake8 = "^6.1.0"
flake8-pyproject = "^1.2.3"
sphinx = "^7.2.6"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.project]
name = "fds"

[tool.pytest.ini_options]
# addopts = [
#     "--import-mode=importlib",
#]
# NOTE: Prevents the ape plugin from activating on our tests
# addopts = "-p no:ape_test"


[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
line_length = 100
force_grid_wrap = 0
include_trailing_comma = true
multi_line_output = 3
use_parentheses = true

[tool.mdformat]
number = true

[[tool.mypy.overrides]]
module = "eth_keys.*"
ignore_missing_imports = true

# * not working so defined in setup.cfg
# [tool.flake8]
# line-length = 100
# exclude = ["__pycache__", ".venv/*"]
